scorer/app/requirements.txt
praw
newspaper3k
twisted